pyproject.toml
[project]
name = "ttt"
version = "0.1.0"
description = "Pixel toys for playing around in the terminal"
authors = [
    {name = "Hamza Haiken",email = "tenchi@team2xh.net"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "numpy (>=2.2.2,<3.0.0)",
    "pyaudio (>=0.2.14,<0.3.0)"
]
packages = [
    { include = "ttt", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ttt = "ttt.__main__:ttt"

[tool.poetry.group.dev.dependencies]
fonttools = "^4.55.4"

